refactor(core): `menu` and `system-tray` feature flags
